+++ Recipe Code/create-edit.html
@@ -1,0 +1,95 @@
+<!DOCTYPE html>
+<html lang="en">
+   
+    <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, intial-scale=1.0" />
+        <meta name="description" content="Recipe storage" />
+        <style>
+            @import url('https://fonts.googleapis.com/css2?family=Open+Sans&display=swap');
+        </style>
+        <link rel="stylesheet" href="style.css">
+        <script src="../testing/frontend src/frontend.js" type="module"></script>
+        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+        <script src="assets/recipeCard.js" type="module"></script>
+        <script src = "editPage.js" type = "module"></script>
+    </head>
+    <body>
+        <header>
+            <section id="main-bar">
+                <a href="user.html">
+                    <img id="grouplogo" src="./grouplogo.PNG" alt="Group Logo">
+                    <h1 id="site-name">App Name</h1>
+                </a>
+            </section>
+        </header>
+        <main>
+            <form id="recipe-create">
+                <section id="input-tags" class="float-right">
+                    <h2 id="input-tags-header" class="line-spacing">Tags</h2>
+                    <div id="tag-input" class="line-spacing">
+                        <input type="text" id="tag-name-input"class="input-elements" placeholder="Enter Tag">
+                        <img id="remove-tag" class="edit-icons input-spacing" src="./remove.PNG" alt="Remove">
+                    </div>
+                    <h3 id="add-tag" class="add-element line-spacing">+ Add Another Tag</h3>
+                </section>
+                <div id="insert-name">
+                    <input type="text" id="input-recipe-name" class="line-spacing" placeholder="Recipe Name">
+                    <span id="name-error"> </span>
+                </div>
+                <div id="insert-image" class="line-spacing">
+                    <img id="file-preview" alt="Insert an image for this recipe.">
+                </div>
+                <input type="file" id="input-file" class="line-spacing">
+                <section id="input-ingredients">
+                    <h2 id="input-ingredients-header" class="line-spacing">Ingredients</h2>
+                    <div id="ingredient-input" class="line-spacing">
+                        <input type="text" id="name-input" class="input-elements input-spacing" placeholder="Ingredient Name">
+                        <input type="text" id="amount-input" class="input-elements input-spacing" placeholder="Amount">
+                        <input type="text" id="unit-input" class="input-elements input-spacing" placeholder="Unit">
+                        <img id="remove-ingredient" class="edit-icons input-spacing" src="./remove.PNG" alt="Remove">
+                    </div>
+                    <h3 id="add-ingredient" class="add-element line-spacing">+ Add Another Ingredient</h3>
+                </section>
+                <section id="input-instructions">
+                    <h2 id="input-instructions-header" class="line-spacing">Instructions</h2>
+                    <ol id="instructions-input-list" class="line-spacing">
+                        <li class="step-input">
+                            <textarea id="step-input-box" rows="2" cols="50" placeholder="Enter instruction step here."></textarea>
+                            <img id="remove-step" class="input-spacing" src="./remove.PNG" alt="Remove">
+                        </li>
+                    </ol>
+                    <h3 id="add-instruction-step" class="add-element line-spacing">+ Add Another Step</h3>
+                </section>
+                
+            </form>
+            <section id="create-edit-buttons">
+                <a href="user.html"><button id="save-recipe" class="float-right create-edit-button">Save</button></a>
+                <a href="user.html"><button id="cancel-recipe" class="float-right create-edit-button">Cancel</button></a>
+            </section>
+        </main>
+        <script>
+            var fileTag = document.getElementById("input-file"),
+            preview = document.getElementById("file-preview");
+    
+            fileTag.addEventListener("change", function() {
+                changeImage(this);
+            });
+
+            function changeImage(input) {
+            var reader;
+
+            if (input.files && input.files[0]) {
+            reader = new FileReader();
+
+            reader.onload = function(e) {
+            preview.setAttribute('src', e.target.result);
+            }
+
+            reader.readAsDataURL(input.files[0]);
+            }
+            }   
+          </script>
+    </body>
+</html>
+
