+++ Recipe Code/index.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, intial-scale=1.0" />
+        <meta name="description" content="Recipe storage" />
+        <style>
+            @import url('https://fonts.googleapis.com/css2?family=Open+Sans&display=swap');
+        </style>
+        <link rel="stylesheet" href="style.css">
+        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+    </head>
+    <body>
+        <header>
+            <section id="main-bar">
+                <a href="index.html">
+                    <img id="grouplogo" src="./grouplogo.PNG" alt="Group Logo">
+                    <h1 id="site-name">App Name</h1>
+                </a>
+            </section>
+        </header>
+        <main>
+            <section id="featured-recipe">
+                <h2 id="recipe-otd-title">Recipe of the Day</h2>
+                
+            </section>
+            <a href="recommended.html"><button id="home-recommended-button" class="float-left home-button">Recommended</button></a>
+            <a href="user.html"><button id="home-recipes-button" class="float-right home-button">My Recipes</button></a>
+        </main>
+    </body>
+</html>
