--- RecipeCode/create-edit.html
@@ -1,74 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-   
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, intial-scale=1.0" />
-        <meta name="description" content="Recipe storage" />
-        <meta name="referrer" content="no-referrer">
-        <style>
-            @import url('https://fonts.googleapis.com/css2?family=Open+Sans&display=swap');
-        </style>
-        <link rel="stylesheet" href="style.css">
-        <script src="./testing/frontend src/frontend.js" type="module"></script>
-        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-        <script src="./assets/recipeCard.js" type="module"></script>
-        <script src = "./editPage.js" type = "module"></script>
-    </head>
-    <body>
-        <header>
-            <section id="main-bar">
-                <a href="./user.html">
-                    <img id="grouplogo" src="./grouplogo.PNG" alt="Group Logo">
-                    <h1 id="site-name">Recipe Chamber</h1>
-                </a>
-            </section>
-        </header>
-        <main>
-            <form id="recipe-create">
-                <section id="input-tags" class="float-right">
-                    <h2 id="input-tags-header" class="line-spacing">Tags</h2>
-                    <article id="tags-inputted"></article>
-                    <h3 id="add-tag" class="add-element line-spacing">+ Add Another Tag</h3>
-                </section>
-                <div id="insert-name">
-                    <input type="text" id="input-recipe-name" class="line-spacing" placeholder="Recipe Name">
-                    <span id="name-error"></span>
-                </div>
-                <img id="file-preview" class="line-spacing" alt="Insert an image for this recipe.">
-                <br>
-                <input type="file" id="input-file" class="line-spacing">
-                <section id="input-ingredients">
-                    <h2 id="input-ingredients-header" class="line-spacing">Ingredients</h2>
-                    <div id="servings">
-                        <label for="servings">Number of Servings:</label>
-                            <select id="serving-number" name="serving-number">
-                                <option value="1">1</option>
-                                <option value="2">2</option>
-                                <option value="3">3</option>
-                                <option value="4">4</option>
-                                <option value="5">5</option>
-                                <option value="6">6</option>
-                                <option value="7">7</option>
-                                <option value="8">8</option>
-                                <option value="9">9</option>
-                                <option value="10">10</option>
-                            </select>
-                    </div>
-                    <article id="ingredients-inputted" class="line-spacing"></article>
-                    <h3 id="add-ingredient" class="add-element line-spacing">+ Add Another Ingredient</h3>
-                </section>
-                <section id="input-instructions">
-                    <h2 id="input-instructions-header" class="line-spacing">Instructions</h2>
-                    <ol id="instructions-input-list" class="line-spacing"></ol>
-                    <h3 id="add-instruction-step" class="add-element line-spacing">+ Add Another Step</h3>
-                </section>
-            </form>
-            <section id="create-edit-buttons">
-                <a href="./user.html"><button id="save-recipe" class="float-right create-edit-button">Save</button></a>
-                <a href="./user.html"><button id="cancel-recipe" class="float-right create-edit-button">Cancel</button></a>
-            </section>
-        </main>
-    </body>
-</html>
-
