--- RecipeCode/recommended.html
@@ -1,64 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, intial-scale=1.0" />
-        <meta name="description" content="Recipe storage" />
-        <style>
-            @import url('https://fonts.googleapis.com/css2?family=Open+Sans&display=swap');
-        </style>
-        <link rel="stylesheet" href="style.css">
-        <script src="./testing/frontend src/frontend.js" type="module"></script>
-        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-        <script src="./assets/recipeCard.js" type="module"></script>
-    </head>
-    <body>
-        <header>
-            <section id="main-bar">
-                <a href="./index.html">
-                    <img id="grouplogo" src="./grouplogo.PNG" alt="Group Logo">
-                    <h1 id="site-name">Recipe Chamber</h1>
-                </a>
-                
-                <a href="./user.html"><button id="my-recipes" class="float-right">My Recipes</button></a>
-                
-                <script>
-                    function recommendedPopUp() {
-                        document.getElementById("popups").classList.add("show");
-                        var test_popup = document.getElementById("recommended-recipe-popup");
-                        test_popup.classList.add("show");
-                    }
-                    function pClose() {
-                        var test_popup = document.getElementById("recommended-recipe-popup");
-                        test_popup.classList.remove("show");
-                        document.getElementById("popups").classList.remove("show");
-                    }
-                </script>
-            </section>
-        </header>
-        <main>
-            <section id="recipe-page-info">
-                <article id="recommended-search-box" class="move-lower float-right">
-                    <input type="text" id="recommended-search-bar" class="search-bar" placeholder="Search">
-                    <button type="button" id="search-button">Search More</button>
-                </article>
-                <h1 id="page-title">Recommended</h1>
-            </section>
-            <section id="recommended-list" class="list"></section>
-            <a href="./index.html"><button id="back-button"><span>&#8592</span> Back</button></a>
-        </main>
-        <div id="popups">
-            <section id="recommended-recipe-popup" class="popup recipe-pages">
-                <button id="recommended-recipe-close" class="float-right line-spacing close-button" onclick="pClose()">X</button>
-                <h2 id="recipe-name" class="line-spacing">Recipe Name</h2>
-                <img id="recipe-image" class="line-spacing" alt="Image of Dish"> <br>
-                <h3 id="ingredients-header" class="line-spacing">Ingredients</h3>
-                <span>Number of servings: <span id="serving-value">1</span></span>
-                <ul id="ingredients" class="line-spacing"></ul>
-                <h3 id="instructions-header" class="line-spacing">Instructions</h3>
-                <ol id="instructions" class="line-spacing"></ol>
-                <button id="add-recipe" class="line-spacing recipe-button">+ Add to my own list</button>
-            </section>
-        </div>
-    </body>
-</html>
