--- RecipeCode/testing/backend src/Test Code/index.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-</head>
-
-<body>
-
-    <div>
-        <button class="clear" type="button">Clear localStorage</button>
-        </br></br>
-        <form>
-            Recipe Name: <input class="name" type='text'><br />
-            Ingredients: <textarea class="ingredients"></textarea><br />
-            Steps: <textarea class="steps"></textarea><br />
-            Tags: <input class="tags" type='text'><br />
-            Image: <input type="file" id='upload' accept="image/*"><br />
-            <button class="button" type="button">Save</button>
-        </form>
-        </br></br>
-        Search: <input class="query" type='text'><br />
-        </br>
-        <div class="text">
-
-        </div>
-</body>
-
-</html>
-<script src="main.js"></script>
