--- RecipeCode/testing/frontend src/README.md
@@ -1,7 +1,0 @@
-Import backend.js
-Create button listeners for all buttons
-
-For save:
-    - create recipe object (instructions in backend.js) and call save(recipe)
-    - use ingredients.js as see fit (ingArray() to be completed)
-    - pls help figure out how to export var imgURL to frontend.js to save
