--- RecipeCode/user.html
@@ -1,133 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, intial-scale=1.0" />
-        <meta name="description" content="Recipe storage" />
-        <meta name="referrer" content="no-referrer">
-        <style>
-            @import url('https://fonts.googleapis.com/css2?family=Open+Sans&display=swap');
-        </style>
-        <link rel="stylesheet" href="style.css">
-        <script src="./testing/frontend src/frontend.js" type="module"></script>
-        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-        <script src="./assets/recipeCard.js" type="module"></script>
-        <script>
-            function popOpen(){
-                document.getElementById("popups").classList.add("show");
-                document.getElementById('grocery-list-popup').classList.add('show');
-            }
-            function popClose(){
-                document.getElementById('grocery-list-popup').classList.remove('show');
-                document.getElementById("popups").classList.remove("show");
-                //added so grocery list could immediately update
-                location.reload();
-            }
-            function recipePopUp() {
-            document.getElementById("popups").classList.add("show");
-            document.getElementById("recipe-popup").classList.add("show");
-            }
-            function testClose() {
-                document.getElementById("recipe-popup").classList.remove("show");
-                document.getElementById("popups").classList.remove("show");
-                //added so grocery list could immediately update
-                location.reload();
-            }
-            function sf() {
-                document.getElementById("popups").classList.add("show");
-                var test_popup = document.getElementById("sort-filter-popup");
-                test_popup.classList.add("show");
-            }
-            function sfClose() {
-                var test_popup = document.getElementById("sort-filter-popup");
-                test_popup.classList.remove("show");
-                document.getElementById("popups").classList.remove("show");
-            }
-            function deleteClose() {
-                var test_popup = document.getElementById("delete-confirmation");
-                test_popup.classList.remove("show");
-            }
-            function deleteYes() {
-                deleteClose();
-                testClose();
-            }
-        </script>
-    </head>
-    <body>
-        <header>
-            <section id="main-bar">
-                <a href="./index.html"><img id="grouplogo" src="./grouplogo.PNG" alt="Group Logo"><h1 id="site-name">Recipe Chamber</h1></a>
-                <a href="./recommended.html"><button id="recommended-recipes" class="float-right">Recommended Recipes</button></a>
-                <button id="grocery-list-button" class="float-right" onclick="popOpen()">Grocery list</button>
-            </section>
-        </header>
-        <main>
-            <section id="recipe-page-info">
-                <article id="action-box" class="move-lower float-right">
-                    <input type="text" id="search-bar" class="search-bar" placeholder="Search">
-                    <button id="sort-button" class="margin-right" onclick="sf()">Sort</button>
-                    <button id="clear-sort" class="margin-right">Reset</button>
-                    <a href="./create-edit.html"><button id="add-button">Add</button></a>
-                </article>
-                <h1 id="page-title">My Recipes</h1>
-            </section>
-            <section id="recipe-list" class="list"></section>
-            <a href="./index.html"><button id="back-button"><span>&#8592</span> Back</button></a>
-        </main>
-        
-    <div id="popups">
-        <section id='grocery-list-popup' class="popup">
-            <button id="grocery-list-close" class="float-right close-button" onclick="popClose()"> X </button>
-            <h2 id="grocery-list-header">Grocery List</h2>
-            <section id="grocery-list-items"></section>
-        </section>
-        <section id="recipe-popup" class="popup recipe-pages">
-            <button type="button" id="recipe-close" class="float-right line-spacing close-button"
-                onclick="testClose()">X</button>
-            <button type="button" id="edit-recipe" class="float-right recipe-button">Edit Recipe</button>
-            <button type="button" id="delete-recipe" class="float-right recipe-button">Delete Recipe</button>
-            <h3 id="tags-header" class="float-right line-spacing">Tags</h3>
-            <div id="recipe-tags" class="float-right"></div>
-            <h2 id="recipe-name" class="line-spacing">Recipe Name</h2>
-            <p id="tracker" class="line-spacing">You have made this <span id="tracker-count">X</span> times, most recent
-                was on <span id="tracker-date">MM/DD/YYYY</span>.</p>
-            <img id="recipe-image" class="line-spacing" alt="Image of Dish"> <br>
-            <h3 id="ingredients-header" class="line-spacing">Ingredients</h3>
-            <span> <input id="slider1" type="range" min="0" max="4" step="1"> <span id='changenum'>original</span> 
-            | Number of servings: <span id="serving-value"></span></span>
-            <ul id="ingredients" class="line-spacing"></ul>
-            <button id="add-grocery" class="line-spacing recipe-button">Add to Grocery List</button>
-            <h3 id="instructions-header" class="line-spacing">Instructions</h3>
-            <ol id="instructions" class="line-spacing"></ol>
-            <button id="track" class="line-spacing recipe-button">Just Made</button>
-        </section>
-        <section id="sort-filter-popup" class="popup">
-            <button id="sort-filter-close" class="float-right line-spacing close-button" onclick="sfClose()">X</button>
-            <h2 id="sort-by-header" class="line-spacing">Sort Recipes By:</h2>
-            <div id="sort-by">
-                <input type="radio" id="most-made" name="sort"><label for="most-made" class="sort"> Most Made </label>
-                <input type="radio" id="least-made" name="sort"><label for="least-made" class="sort"> Least Made
-                </label>
-                <input type="radio" id="alphabetical" name="sort" checked><label for="alphabetical" class="sort">
-                    Alphabetical Order (A-Z) </label>
-                <input type="radio" id="newest" name="sort"><label for="newest" class="sort"> Newest to Oldest </label>
-                <input type="radio" id="oldest" name="sort"><label for="oldest" class="sort"> Oldest to Newest </label>
-            </div>
-            <hr>
-            <h2 id="filter-tags" class="line-spacing">Filter By Tags:</h2>
-            <div id="tags"></div>
-        </section>
-    </div>
-    <section id="delete-confirmation">
-        <div id="delete-box">
-            <p>Are you sure you want to delete '<span id="delete-recipe-name">Recipe Name</span>'?</p>
-            <br>
-            <button type="button" id="delete-yes" class="recipe-button delete-button" onclick="deleteYes()"> Yes
-            </button>
-            <button type="button" id="delete-no" class="recipe-button delete-button" onclick="deleteClose()"> No
-            </button>
-        </div>
-    </section>
-</body>
-
-</html>
