+++ package.json
@@ -4,7 +4,6 @@
   "description": "[Our Team Page](admin/team.md)",
   "main": "index.js",
   "scripts": {
-    "build": "",
     "doc": "jsdoc -c jsdoc.json",
     "test": "jest"
   },
